+++ noodles-bcf/src/async/io/writer/header.rs
@@ -1,15 +1,14 @@
-use std::{
-    ffi::CString,
-    io::{self, Write},
-};
+use std::ffi::CString;
 
-use byteorder::{LittleEndian, WriteBytesExt};
 use noodles_vcf as vcf;
+use tokio::io::{self, AsyncWrite, AsyncWriteExt};
 
-pub(super) fn write_header<W>(writer: &mut W, header: &vcf::Header) -> io::Result<()>
+pub(super) async fn write_header<W>(writer: &mut W, header: &vcf::Header) -> io::Result<()>
 where
-    W: Write,
+    W: AsyncWrite + Unpin,
 {
+    use crate::io::writer::header::serialize_header;
+
     let raw_header = serialize_header(header)?;
     let c_raw_header =
         CString::new(raw_header).map_err(|e| io::Error::new(io::ErrorKind::InvalidInput, e))?;
@@ -18,27 +17,21 @@ where
     let l_text =
         u32::try_from(text.len()).map_err(|e| io::Error::new(io::ErrorKind::InvalidInput, e))?;
 
-    writer.write_u32::<LittleEndian>(l_text)?;
-    writer.write_all(text)?;
+    writer.write_u32_le(l_text).await?;
+    writer.write_all(text).await?;
 
     Ok(())
-}
-
-pub(crate) fn serialize_header(header: &vcf::Header) -> io::Result<Vec<u8>> {
-    let mut writer = vcf::io::Writer::new(Vec::new());
-    writer.write_header(header)?;
-    Ok(writer.into_inner())
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
 
-    #[test]
-    fn test_write_header() -> io::Result<()> {
+    #[tokio::test]
+    async fn test_write_header() -> io::Result<()> {
         let mut buf = Vec::new();
         let header = vcf::Header::default();
-        write_header(&mut buf, &header)?;
+        write_header(&mut buf, &header).await?;
 
         let mut expected = 61i32.to_le_bytes().to_vec();