+++ noodles-bcf/src/lib.rs
@@ -11,6 +11,6 @@ pub mod record;
 pub use self::record::Record;
 
 #[cfg(feature = "async")]
-pub use self::r#async::io::Reader as AsyncReader;
+pub use self::r#async::io::{Reader as AsyncReader, Writer as AsyncWriter};
 
 static MAGIC_NUMBER: &[u8] = b"BCF";